+++ apps/app/lib/router/routes/home_route.dart
@@ -5,8 +5,8 @@ import 'package:cores_navigation/providers.dart';
 import 'package:features_debug_mode/ui.dart';
 import 'package:features_webview/webview.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_app/feature/home/ui/home_navigator.dart';
 import 'package:flutter_app/feature/home/ui/home_page.dart';
+import 'package:flutter_app/router/navigator/home_navigator.dart';
 import 'package:flutter_app/router/routes/setting_route.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
